LayoutTests/fast/frames/sandboxed-iframe-close-top-noclose-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate a navigation change for frame with URL close-top.html.

This test verifies that a sandboxed IFrame can close a top-level frame with allow-top-navigation.

Start Test
